+++ django_component_kit/__init__.py
@@ -1,13 +1,15 @@
-from .tags import do_merge_attrs, do_slot, do_render_slot, do_partial
+from .tags import do_merge_attrs, do_slot, do_render_slot, do_partial, do_assets
 from .decorators import component_inline_tag, component_block_tag
-from .partials import render_partial_from_template
+from .partials import render_partial_from_template, PartialResponse
 
 __all__ = [
     "do_merge_attrs",
     "do_slot",
     "do_render_slot",
+    "do_assets",
     "component_inline_tag",
     "component_block_tag",
     "do_partial",
     "render_partial_from_template",
+    "PartialResponse",
 ]